test: update utils tests
